Update refetch thumbnail API to process custom thumbnails
GET request = tries to fetch thumbnail from the og:image inside <meta>
tag from the bookmarked website itself

POST request = user can upload their own image as a thumbnail. There are
no checks to determine if the uploaded file is actually an image, and
there will never be one. But, in the function located at
@internal/server/template_functions.go > func byteConversion
you'll find the image types that is rendered in the web UI.

Maximum file size is set to ~5,23MB.

An example of the POST request using curl:
curl -X POST \
    -F "thumbnail=@/tmp/test.jpg" \
    http://localhost:41415/api/refetch-thumbnail/1748
thumbnail = form key that needs to have the thumbnail image.
1748 = bookmark's ID to update thumbnail for.

Can use this to make a bash script to upload custom thumbnail for sites
that do not have og:image. At least that has been my use case.
Also, uploading a specific thumbnail for select bookmarks to be able to
easily spot them.